+++ crates/orchestrator/src/cli/mod.rs
@@ -613,6 +613,13 @@ pub mod validate_params {
                 if s.is_empty() { None } else { Some(s.parse::<u64>().expect("Failed to parse min block to process")) }
             }),
             service_id: service_args.service_id.clone().and_then(|s| if s.is_empty() { None } else { Some(s) }),
+            max_concurrent_snos_jobs: service_args.max_concurrent_snos_jobs.clone().and_then(|s| {
+                if s.is_empty() {
+                    None
+                } else {
+                    Some(s.parse::<u64>().expect("Failed to parse max concurrent SNOS jobs"))
+                }
+            }),
         })
     }
 
@@ -894,6 +901,7 @@ pub mod validate_params {
                 max_block_to_process: Some("66645".to_string()),
                 min_block_to_process: Some("100".to_string()),
                 service_id: Some("orchestrator_1".to_string()),
+                max_concurrent_snos_jobs: Some("10".to_string()),
             };
             let service_params = validate_service_params(&service_args);
             assert!(service_params.is_ok());
@@ -901,6 +909,7 @@ pub mod validate_params {
             assert_eq!(service_params.max_block_to_process, Some(66645));
             assert_eq!(service_params.min_block_to_process, Some(100));
             assert_eq!(service_params.service_id, Some("orchestrator_1".to_string()));
+            assert_eq!(service_params.max_concurrent_snos_jobs, Some(10));
         }
     }
 }