+++ crates/orchestrator/src/tests/config.rs
@@ -545,10 +545,19 @@ fn get_env_params() -> EnvParams {
     let min_block: Option<u64> = env.and_then(|s| if s.is_empty() { None } else { Some(s.parse::<u64>().unwrap()) });
 
     let env = get_env_var_optional("MADARA_ORCHESTRATOR_SERVICE_ID").expect("Couldn't get service id");
-    let service_id: Option<String> =
-        env.and_then(|s| if s.is_empty() { None } else { Some(s.parse::<String>().unwrap()) });
+    let service_id: Option<String> = env.and_then(|s| if s.is_empty() { None } else { Some(s) });
 
-    let service_config = ServiceParams { max_block_to_process: max_block, min_block_to_process: min_block, service_id };
+    let env = get_env_var_optional("MADARA_ORCHESTRATOR_MAX_CONCURRENT_SNOS_JOBS")
+        .expect("Couldn't get max concurrent snos jobs");
+    let max_concurrent_snos_jobs: Option<u64> =
+        env.and_then(|s| if s.is_empty() { None } else { Some(s.parse::<u64>().unwrap()) });
+
+    let service_config = ServiceParams {
+        max_block_to_process: max_block,
+        min_block_to_process: min_block,
+        service_id,
+        max_concurrent_snos_jobs,
+    };
 
     let server_config = ServerParams {
         host: get_env_var_or_panic("MADARA_ORCHESTRATOR_HOST"),